+++ CHANGELOG.md
@@ -14,13 +14,19 @@ DONE
 - ProfilePage
   - Multiple majors / minors
 
+- Side Navigation Bar
+  - Will be the menu
+
 - CSS Styling
   - ProfilePage
   - SettingsPage
   - Signup 
   - Login
   - Dashboard
+  - Sidebar
+
+App.js
+  - If user is not signed in, dashboard/profile/settings should redirect to /login
 
 TODO
-  - If user is not signed in, /dashboard should redirect to /login
   - ISSUE(1): Let new user take username of a deleted account when (1) signing up (2)choosing new username in settings